pyproject.toml
[project]
name = "api-gateway-ninja"
version = "0.1.0"
description = ""
authors = [
    {name = "Anderson Bairo",email = "anderson.bairo@yahoo.com.br"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-ninja (>=1.4.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "aiohttp (>=3.11.16,<4.0.0)",
    "parse (>=1.20.2,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
